Upgrade code and get it all working once again
This fixes year long issues in regards to linking. It's been frustrating
to say the least and was a big reason I let the code bit rot. Given
their is a pandemic goin on now and well I got bored I took the time to
fix all the code. This moves it into Rust 2018 edition works on a newer
version of GHC and overall reformats and cleans up the code with some
better libraries. This isn't the state the release will be in, but it's
just nice to have all the test suites working again and maybe now I can
make some actual progress on creating a good library.

+++ src/types.rs
@@ -1,10 +1,6 @@
 //! Contains Type Equivalency for Haskell Types
 
-pub use std::os::raw::{
-	c_char,
-	c_float,
-	c_double
-};
+pub use std::os::raw::{c_char, c_double, c_float};
 
 /// FFI Version of bool
 pub type Boolean = u8;
@@ -16,7 +12,7 @@ pub type Chr = c_char;
 pub type Str = *const c_char;
 
 /// FFI version of u8
-pub type U8  = u8;
+pub type U8 = u8;
 
 /// FFI version of u16
 pub type U16 = u16;
@@ -28,7 +24,7 @@ pub type U32 = u32;
 pub type U64 = u64;
 
 /// FFI version of i8
-pub type I8  = i8;
+pub type I8 = i8;
 
 /// FFI version of i16
 pub type I16 = i16;
